Extract session image loading pipeline
Move still image decoding, animated frame loading, cache usage, dtype normalization,
and render preparation out of session_window.py into SessionImageLoaderMixin.
Wire SessionDisplay to compose the new mixin.
